requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
httpx>=0.28.0
tiktoken>=0.8.0
pydantic>=2.10.0
jinja2>=3.1.0
python-multipart>=0.0.18
aiofiles>=24.1.0
pyyaml>=6.0.0
itsdangerous>=2.2.0
